requirements.txt
Flask == 1.1.1
Flask_Cors == 3.0.8
Flask_RESTful == 0.3.7
PyYAML == 5.3
aether == 0.3.37
awake-analysis-tools == 0.0.2
cronyo == 0.4.2
data-config-controller == 0.0.3
django-utilities == 0.1.5
dropbox == 9.4.0
fan-duel == 0.1
general-utilities == 0.1.1
ih-common == 1.3.10
lama-phenotype-detection == 0.9.2
meross-iot == 0.3.2.20
pandas == 0.25.1
pricingengine == 0.0.5-beta-
psopy == 0.2.3
pymata-aio == 2.35
rapydo-utils == 1.0.0
redsolutioncms-django-utilities == 0.1.5
requests == 2.18.4
sibl-gui == 4.0.8
snake-on-pygame == 1.0
troppo == 0.0.4
volkswagencarnet == 4.1.5
weatherapplication == 0.0.2
weatherapplication-arorarajat084 == 0.0.2
weatherapplication-rajat == 0.0.1
weatherapplicationarorarajat == 0.0.4
yoc == 0.0.0